requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
click==8.3.2
colorama==0.4.6
ecdsa==0.19.2
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Mako==1.3.11
MarkupSafe==3.0.3
packaging==26.1
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.3
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
psycopg==3.3.3
psycopg-binary==3.3.3
pytest==9.0.3
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.26
reportlab==4.2.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
tzdata==2025.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
